services/whisper-livekit/requirements.txt
openai-whisper>=20231117
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
soundfile>=0.12.0
